Cast to uint16_t explicitly as struct CommandBuffer::Statistics has fields of this type

+++ src/core/macros.h
@@ -72,7 +72,7 @@ along with this program. If not, see <https://www.gnu.org/licenses/>.
 #endif
 
 #ifdef MAGMA_STATS
-    #define MAGMA_INCR(cnt, n) cnt += n
+    #define MAGMA_INCR(cnt, n) cnt += (uint16_t)n
 #else
     #define MAGMA_INCR(cnt, n)
 #endif // MAGMA_STATS
